abstract coco backend from mAP
